Added Class Diagrams to Repository
I have added a class diagram representing the objects we discussed in class today -
Hopefully this will be a good talking point and something we could include in this week's blog. 
I have made a couple of changes, so please review and let me know if you would like anything to be changed.